# specs/001-physical-ai-textbook/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Physical AI Textbook API
  description: |
    Backend API for the Physical AI & Humanoid Robotics Textbook platform.
    Provides authentication, RAG-powered chat, and progress tracking.
  version: 1.0.0
  contact:
    name: Panaversity
    url: https://panaversity.com
servers:
  - url: https://api.physical-ai-textbook.example.com
    description: Production server
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/chat/query:
    post:
      tags: [chat]
      summary: Query the RAG chatbot
      description: |
        Send a question to the chatbot. Optionally include selected text
        for context-aware responses. Returns complete response.
      operationId: queryChat
      security:
        - bearerAuth: []
        - {}  # Also allows unauthenticated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatQuery'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chat/query/stream:
    post:
      tags: [chat]
      summary: Query chatbot with streaming response
      description: |
        Same as /query but returns Server-Sent Events for real-time streaming.
        First event contains sources, subsequent events contain response chunks.
      operationId: queryChatStream
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatQuery'
      responses:
        '200':
          description: Streaming response
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE stream with events:
                  - type: "sources" - Retrieved context chunks
                  - type: "chunk" - Response text chunk
                  - type: "done" - Stream complete
  /api/chat/sessions:
    get:
      tags: [chat]
      summary: List user's chat sessions
      operationId: listChatSessions
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Chat sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSessionSummary'
  /api/chat/sessions/{session_id}:
    get:
      tags: [chat]
      summary: Get chat session with messages
      operationId: getChatSession
      security:
        - bearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before_sequence
          in: query
          description: Pagination - get messages before this sequence number
          schema:
            type: integer
      responses:
        '200':
          description: Chat session with messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetail'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [chat]
      summary: Archive a chat session
      operationId: archiveChatSession
      security:
        - bearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session archived
        '404':
          description: Session not found
  /api/chat/messages/{message_id}/feedback:
    post:
      tags: [chat]
      summary: Submit feedback for a message
      operationId: submitMessageFeedback
      security:
        - bearerAuth: []
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageFeedback'
      responses:
        '204':
          description: Feedback recorded
        '404':
          description: Message not found

  # ============ Progress ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable message
        details:
          type: object
          additionalProperties: true

    # ============ Auth ============
...
    ChatQuery:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
        selected_text:
          type: string
          maxLength: 5000
          description: Text selected by user for context
        module_id:
          type: string
          description: Filter retrieval to specific module
        chapter_id:
          type: string
          description: Filter retrieval to specific chapter
        session_id:
          type: string
          format: uuid
          description: Continue existing session
        chat_history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
    ChatResponse:
      type: object
      required: [answer, sources, session_id]
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/ChatSource'
        session_id:
          type: string
          format: uuid
    ChatSource:
      type: object
      required: [module_id, chapter_id, section, score]
      properties:
        module_id:
          type: string
        chapter_id:
          type: string
        section:
          type: string
        score:
          type: number
          format: float
    ChatSessionSummary:
      type: object
      required: [id, message_count, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        message_count:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChatSessionDetail:
      type: object
      required: [id, messages]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        initial_context:
          type: object
          nullable: true
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
    ChatMessage:
      type: object
      required: [id, sequence_number, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        sequence_number:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        retrieved_chunks:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ChatSource'
        selection_context:
          type: object
          nullable: true
          properties:
            selected_text:
              type: string
            chapter_id:
              type: string
        feedback_rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        created_at:
          type: string
          format: date-time
    MessageFeedback:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedback_text:
          type: string
          maxLength: 1000

    # ============ Progress ============
...